* Fixed time_t vs. long warnings (typecasted properly as '(long)').
git-svn-id: http://svn.ratbox.org/svnroot/trunk@7391 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/s_conf.c
@@ -2246,14 +2246,14 @@ void WriteKlineOrDline( KlineType type,
                reason,
 	       current_date,
 	       get_oper_name(source_p),
-               cur_time);
+               (long) cur_time);
   else
     ircsprintf(buffer, "\"%s\",\"%s %s\",\"%s\",%ld\n",
                host,
                reason,
 	       current_date,
 	       get_oper_name(source_p),
-               cur_time);
+               (long) cur_time);
 
 
   if (fbputs(buffer,out) == -1)
